refactor(core): split stateless strategy into function modules
