+++ packages/core/src/session/stateful/createSession.ts
@@ -1,6 +1,6 @@
 import { AuraAuthError } from "@/shared/errors.ts"
 import { createHash, createSecretValue } from "@/shared/crypto.ts"
-import { createDevice as __createDevice } from "@/session/stateful/utils.ts"
+import { createDevice as __createDevice } from "@/shared/utils/session-strategy.ts"
 import type { TypedJWTPayload } from "@aura-stack/jose"
 import type { InternalStatefulContext, User } from "@/@types/index.ts"
