+++ packages/core/src/session/stateful/oauthCallback.ts
@@ -1,13 +1,13 @@
-import { isRelativeURL, isSameOrigin, isTrustedOrigin } from "@/shared/assert.ts"
-import { createCSRF, createHash, createSecretValue } from "@/shared/crypto.ts"
+import { HeadersBuilder } from "@aura-stack/router"
 import { AuraAuthError } from "@/shared/errors.ts"
 import { validateIDToken } from "@/shared/oidc/id-token.ts"
-import { isOIDCProvider, resolveOpenIDProvider } from "@/shared/oidc/resolve-provider.ts"
-import { getOriginURL, getTrustedOrigins } from "@/shared/utils/authorization.ts"
 import { createAccessToken, getUserInfo } from "@/shared/utils/oauth.ts"
+import { createCSRF, createHash, createSecretValue } from "@/shared/crypto.ts"
+import { getOriginURL, getTrustedOrigins } from "@/shared/utils/authorization.ts"
+import { isRelativeURL, isSameOrigin, isTrustedOrigin } from "@/shared/assert.ts"
+import { createDevice as __createDevice } from "@/shared/utils/session-strategy.ts"
+import { isOIDCProvider, resolveOpenIDProvider } from "@/shared/oidc/resolve-provider.ts"
 import type { InternalStatefulContext } from "@/@types/session.ts"
-import { createDevice as __createDevice } from "./utils.ts"
-import { HeadersBuilder } from "@aura-stack/router"
 
 export const __oauthCallback = ({ ctx, cookies, cookieManager }: InternalStatefulContext) => {
     const { logger, jose, oauth, sessionConfig } = ctx
