+++ packages/core/src/session/stateful/revokeToken.ts
@@ -1,8 +1,8 @@
-import { AuraAuthError } from "@/shared/errors.ts"
-import { revokeProviderToken } from "@/shared/utils/revoke-token.ts"
 import { HeadersBuilder } from "@aura-stack/router"
-import { secureApiHeaders } from "@/shared/headers.ts"
 import { getExpiredCookie } from "@/cookie.ts"
+import { AuraAuthError } from "@/shared/errors.ts"
+import { secureApiHeaders } from "@/shared/headers.ts"
+import { revokeProviderToken } from "@/shared/utils/revoke-token.ts"
 import { getErrorName, toUnionHeaders } from "@/shared/utils.ts"
 import type { InternalStatefulContext } from "@/@types/session.ts"
